commonCompileAndCreate
